add unit tests
